Simplify Controller & Abstract Generation
